backend/requirements.txt
# FAKE-SHA Backend Dependencies
# Install with: pip install -r requirements.txt

fastapi==0.115.5
uvicorn[standard]==0.32.1
supabase==2.10.0
python-dotenv==1.0.1
pandas>=2.0
scikit-learn>=1.3.0
# RoBERTa (set FAKE_SHA_ANALYZER=roberta)
torch>=2.1.0
transformers>=4.36.0
safetensors>=0.4.0
datasets>=2.16.0
